+++ lib/elegant_notification.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:elegant_notification/resources/arrays.dart';
 import 'package:elegant_notification/resources/colors.dart';
+import 'package:elegant_notification/resources/constants.dart';
 import 'package:elegant_notification/resources/extensions.dart';
 import 'package:elegant_notification/widgets/animated_progress_bar.dart';
 import 'package:elegant_notification/widgets/toast_content.dart';
@@ -28,7 +29,7 @@ class ElegantNotification extends StatefulWidget {
     this.notificationPosition = NotificationPosition.topRight,
     this.animation = AnimationType.fromRight,
     this.animationDuration = const Duration(milliseconds: 600),
-    this.iconSize = 20,
+    this.iconSize = defaultIconSize,
     this.action,
     this.onActionPressed,
     this.autoDismiss = true,
@@ -38,59 +39,7 @@ class ElegantNotification extends StatefulWidget {
     this.progressIndicatorBackground = greyColor,
   }) : super(key: key) {
     notificationType = NotificationType.custom;
-
-    if (showProgressIndicator) {
-      assert(autoDismiss != false);
-    }
-    if (action != null) {
-      assert(onActionPressed != null);
-    }
-
-    if (notificationPosition == NotificationPosition.centerRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.centerLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.topCenter) {
-      assert(
-        animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromRight,
-      );
-    } else if (notificationPosition == NotificationPosition.topRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromBottom,
-      );
-    } else if (notificationPosition == NotificationPosition.topLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromBottom,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomCenter) {
-      assert(
-        animation != AnimationType.fromTop &&
-            animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromRight,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromTop,
-      );
-    }
+    checkAssertions();
   }
 
   ElegantNotification.success({
@@ -103,7 +52,7 @@ class ElegantNotification extends StatefulWidget {
     this.displayCloseButton = true,
     this.onCloseButtonPressed,
     this.onProgressFinished,
-    this.iconSize = 20,
+    this.iconSize = defaultIconSize,
     this.notificationPosition = NotificationPosition.topRight,
     this.animation = AnimationType.fromRight,
     this.animationDuration = const Duration(milliseconds: 600),
@@ -119,59 +68,7 @@ class ElegantNotification extends StatefulWidget {
     notificationType = NotificationType.success;
     progressIndicatorColor = successColor;
     icon = null;
-
-    if (showProgressIndicator) {
-      assert(autoDismiss != false);
-    }
-    if (action != null) {
-      assert(onActionPressed != null);
-    }
-
-    if (notificationPosition == NotificationPosition.centerRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.centerLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.topCenter) {
-      assert(
-        animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromRight,
-      );
-    } else if (notificationPosition == NotificationPosition.topRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromBottom,
-      );
-    } else if (notificationPosition == NotificationPosition.topLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromBottom,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomCenter) {
-      assert(
-        animation != AnimationType.fromTop &&
-            animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromRight,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromTop,
-      );
-    }
+    checkAssertions();
   }
 
   ElegantNotification.error({
@@ -184,7 +81,7 @@ class ElegantNotification extends StatefulWidget {
     this.displayCloseButton = true,
     this.onCloseButtonPressed,
     this.onProgressFinished,
-    this.iconSize = 20,
+    this.iconSize = defaultIconSize,
     this.notificationPosition = NotificationPosition.topRight,
     this.animation = AnimationType.fromRight,
     this.animationDuration = const Duration(milliseconds: 600),
@@ -200,59 +97,7 @@ class ElegantNotification extends StatefulWidget {
     notificationType = NotificationType.error;
     progressIndicatorColor = errorColor;
     icon = null;
-
-    if (showProgressIndicator) {
-      assert(autoDismiss != false);
-    }
-    if (action != null) {
-      assert(onActionPressed != null);
-    }
-
-    if (notificationPosition == NotificationPosition.centerRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.centerLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.topCenter) {
-      assert(
-        animation != AnimationType.fromBottom &&
-            animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromRight,
-      );
-    } else if (notificationPosition == NotificationPosition.topRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromBottom,
-      );
-    } else if (notificationPosition == NotificationPosition.topLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromBottom,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomCenter) {
-      assert(
-        animation != AnimationType.fromTop &&
-            animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromRight,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomRight) {
-      assert(
-        animation != AnimationType.fromLeft &&
-            animation != AnimationType.fromTop,
-      );
-    } else if (notificationPosition == NotificationPosition.bottomLeft) {
-      assert(
-        animation != AnimationType.fromRight &&
-            animation != AnimationType.fromTop,
-      );
-    }
+    checkAssertions();
   }
 
   ElegantNotification.info({
@@ -265,7 +110,7 @@ class ElegantNotification extends StatefulWidget {
     this.displayCloseButton = true,
     this.onCloseButtonPressed,
     this.onProgressFinished,
-    this.iconSize = 20,
+    this.iconSize = defaultIconSize,
     this.notificationPosition = NotificationPosition.topRight,
     this.animation = AnimationType.fromRight,
     this.animationDuration = const Duration(milliseconds: 600),
@@ -281,7 +126,11 @@ class ElegantNotification extends StatefulWidget {
     notificationType = NotificationType.info;
     progressIndicatorColor = inforColor;
     icon = null;
+    checkAssertions();
+  }
 
+  ///Checks assertions for various constructors of this package
+  void checkAssertions() {
     if (showProgressIndicator) {
       assert(autoDismiss != false);
     }
@@ -347,9 +196,9 @@ class ElegantNotification extends StatefulWidget {
   ///
   late Widget? icon;
 
-  ///The size of the icon, by default it's 40px
+  ///The size of the icon, by default it's 20px
   ///
-  late double iconSize;
+  final double iconSize;
 
   ///The type of the animation set on the notification
   ///possible values